server/requirements.txt
fastapi>=0.135.0,<1.0
uvicorn[standard]>=0.42.0,<1.0
sqlalchemy[asyncio]>=2.0.48,<3.0
asyncpg>=0.31.0,<1.0
alembic>=1.18.0,<2.0
python-dotenv>=1.2.0,<2.0
httpx>=0.27.0,<1.0
pyjwt>=2.12.0,<3.0
passlib[bcrypt]>=1.7.4,<2.0
pydantic-settings>=2.13.0,<3.0
python-multipart>=0.0.9,<1.0
docker>=7.0.0,<8.0
pgvector>=0.3.0,<1.0
openai>=1.0.0,<2.0
beautifulsoup4>=4.12.0,<5.0
jsonschema>=4.21.0,<5.0
